refactor: remove values in new file
