+++ homework-testing/tests/values.js
@@ -1,0 +1,32 @@
+const valuesArray = [
+  {
+    key: 30,
+    value: '30',
+  },
+  {
+    key: 25,
+    value: '25',
+  },
+  {
+    key: 60,
+    value: '60',
+  },
+  {
+    key: 50,
+    value: '50',
+  },
+  {
+    key: 100,
+    value: '100',
+  },
+  {
+    key: 27,
+    value: '27',
+  },
+  {
+    key: 15,
+    value: '15',
+  },
+];
+
+module.exports = valuesArray;
